requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b8
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
emoji==2.2.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
pydantic==1.10.7
pytz==2023.3
typing-extensions==4.5.0
yarl==1.9.2
